+++ tests/luau_files/type_tables/input.lua
@@ -1,0 +1,10 @@
+type PromptSettings = {
+    object: string,
+    action: string,
+    holdDuration: number,
+    keyboardKey: KeyCode,
+    gamepadKey: KeyCode,
+    distance: number,
+    lineOfSight: boolean,
+    offset: Vector2,
+}